Line 28 swapped 'ontouchstart' for 'ontouchend'
On sliders with large previous/next buttons it is impossible to scroll
up and down the page if the user accidentally touches the button area.
'ontouchend' waits for the touch to release thereby confirming the user
actually wants to 'click' the button rather than scroll…a minor change
but gets rid of the annoying problem of the user trying to avoid
touching an area on the screen, when all they want to do is get up and
down the page.