Added new api method `stop()`
It's clear from the code that it does.

Say you want to integrate vimeo or over video player with flexslider,
in example provided with flexslider when you play the video the slider
pauses, pressing pause on video flexslider resumes.

Say you want to do the same but having 'pauseOnHover` set to true, pressing play button on video pauses flexslider, but if you hover out of the flexslider
it resumes sliding, because of hover off event.

Now hoveroff event checks if player wasn't stoped at all before resuming.

I think it's clear why and when you need to use `stop()`