Bug fix for `play` method
Calling multiple times `play` we are issuing multiple timers, but
flexslider is designed to handle only one, now we check if one already
exists and if not we'll not set another timer.

Also say you call `play` method 2 times, and then `pause` it was clearing
only the last interval so slider was working like nothing happened.