Display image with its own height on stage.
If a slider has images of varying heights, the stage height is set to the height of the tallest image. Images that are shorter than the tallest image are displayed from the upper left part of the stage. This leads to extra white space below (and potentially to the side of) the image. This commit fixes this issue for the fade animation.

Prior to this commit, images were hidden and shown by toggling the "z-index" between 1 (hidden) and 2 (shown). The problem with this approach is that each image's "display" property is still "block" and therefore will hold the height of the bounding element. This scenario leads to the tallest image setting the height of the stage.

This commit changes the show/hide strategy. Instead of toggling a "z-index" value, the "display" value is toggled between "block" (shown) and "none" (hidden).

Note that this fix does not yet address issues with the slide animation.

Relevant issues:
woothemes/FlexSlider#426
woothemes/FlexSlider#364
woothemes/FlexSlider#436