Narrowing target's index() in methods.asNav #1213
Within the Thumbnail Slider implementation, you may want to display only some of the slides – i.e. `<li>` elements – throught a customised selector property" –  for instance to set up lightbox galleries of which you'll only show cover pictures in the slider.
At line 164, `target` variable calls for the `index()` function, which looks for the selector among all siblings of the page, without taking into account the parent container – i.e. the container on which flexslider have been initialised (`#slider` and `#carousel` in the official Thumbnail Slider code sample). This causes a bug (see #1213) that seems to be fixed by this commit.