Fix Event Handler for Added Thumb Slides
The click handler was being bound on each slide and when a slide is added,
it wasn't being attached to new slides. Uses event bubbling instead so
there is only one handler on the parent slider and we don't have to
remember to attach the handle when a slide is added.